requirements.txt
addict==2.4.0
affine==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiomultiprocess==0.9.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contextily==1.6.2
contourpy==1.3.0
coord-convert==0.2.1
coverage==7.6.1
crcmod==1.7
cryptography==42.0.5
cycler==0.12.1
dashscope==1.20.10
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.0
filelock==3.16.1
fiona==1.10.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.2.3
gast==0.5.4
GDAL==3.4.1
generate-od==0.1
geographiclib==2.0
geojson==3.1.0
geopandas==0.14.3
geopy==2.4.1
grpcio==1.66.1
grpcio-tools==1.62.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
igraph==0.11.6
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==0.10.0
joblib==1.4.2
json_repair==0.25.3
jsonlines==4.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Levenshtein==0.25.1
loguru==0.7.2
lru-dict==1.3.0
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.7
mercantile==1.2.1
modelscope==1.13.3
mosstool==0.2.34
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
openai==1.46.0
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
parso==0.8.4
path4gmns==0.9.8
peft==0.12.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.0.0
pyarrow-hotfix==0.6
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.0
Pygments==2.18.0
pymongo==4.9.1
pyparsing==3.1.4
pyproj==3.6.1
pytest==8.3.3
pytest-cov==5.0.0
pytest-cover==3.0.0
pytest-coverage==0.0
python-dateutil==2.9.0.post0
python-moss==0.4.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.7
rasterio==1.3.9
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.0
seaborn==0.13.2
setproctitle==1.3.3
shapely==2.0.6
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.6
sse-starlette==2.1.3
stack-data==0.6.3
starlette==0.38.5
stringcase==1.2.0
sympy==1.13.3
tenacity==8.5.0
texttable==1.7.0
thefuzz==0.22.1
threadpoolctl==3.5.0
timm==1.0.9
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.41.2
triton==2.1.0
trl==0.9.6
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
wcwidth==0.2.13
websocket-client==1.8.0
xxhash==3.4.1
xyzservices==2024.9.0
yapf==0.40.2
yarl==1.11.1
zipp==3.18.1
